Support Paging for ListOperation
By default ALL results are loaded. This is the standard in e.g.
command line tools. In an app you can use the paging limit and a
special config of the operation to load each page one by one.

Remove subcommand list from BuildsOperation. Please use the good old 
ListOperation for this!

Add UrlOperation to request a simple URL ready to go with query 
parameters etc.